refactor: refactor `_strings` and `_strconv` to `str`(because private package)
