+++ str/docs.go
@@ -1,4 +1,4 @@
-package _strings
+package str
 
 // use `_strings` instead of `strings` to avoid name collision
 // some functions see https://www.php.net/manual/zh/ref.strings.php
